refactor: put validation logic to validator
